requirements.txt
Flask
Pillow
rembg
opencv-python
matplotlib
numpy
